Move get_list_of_outliers from sequence and into utils
