+++ anvio/sequence.py
@@ -140,74 +140,3 @@ class Composition:
             self.GC_content = (self.G + self.C) * 1.0 / length
 
 
-def get_indices_for_outlier_values(c):
-    is_outlier = get_list_of_outliers(c)
-    return set([p for p in range(0, c.size) if is_outlier[p]])
-
-
-def get_list_of_outliers(values, threshold=None, zeros_are_outliers=False, median=None):
-    """Return boolean array of whether values are outliers (True means yes)
-
-    Modified from Joe Kington's (https://stackoverflow.com/users/325565/joe-kington)
-    implementation computing absolute deviation around the median.
-
-    Parameters
-    ==========
-    values : array-like
-        An numobservations by numdimensions array of observations
-
-    threshold : number, None
-        The modified z-score to use as a thresholdold. Observations with
-        a modified z-score (based on the median absolute deviation) greater
-        than this value will be classified as outliers.
-
-    median : array-like, None
-        Pass median value of values if you already calculated it to save time
-
-    Returns
-    =======
-    mask : numpy array (dtype=bool)
-        A numobservations-length boolean array.
-
-    References
-    ==========
-    Boris Iglewicz and David Hoaglin (1993), "Volume 16: How to Detect and
-    Handle Outliers", The ASQC Basic References in Quality Control:
-    Statistical Techniques, Edward F. Mykytka, Ph.D., Editor.
-
-    http://www.sciencedirect.com/science/article/pii/S0022103113000668
-    """
-
-    if threshold is None:
-        threshold = 1.5
-
-    if len(values.shape) == 1:
-        values = values[:, None]
-
-    if not median: median = np.median(values, axis=0)
-
-    diff = np.sum((values - median) ** 2, axis=-1)
-    diff = np.sqrt(diff)
-    median_absolute_deviation = np.median(diff)
-
-    if not median_absolute_deviation:
-       if values[0] == 0:
-            # A vector of all zeros is considered "all outliers"
-            return np.array([True] * values.size)
-       else:
-            # A vector of uniform non-zero values is "all non-outliers"
-            # This could be important for silly cases (like in megahit) in which there is a maximum value for coverage
-            return np.array([False] * values.size)
-
-    modified_z_score = 0.6745 * diff / median_absolute_deviation
-    non_outliers = modified_z_score > threshold
-
-    if not zeros_are_outliers:
-        return non_outliers
-    else:
-        zero_positions = [x for x in range(len(values)) if values[x] == 0]
-        for i in zero_positions:
-            non_outliers[i] = True
-        return non_outliers
-
-
